fix: sucessfully conect with spectrometer
