+++ frontend/src/app/users-management/page.tsx
@@ -92,11 +92,11 @@ function ManageUsers() {
     control: (provided) => ({
       ...provided,
       borderRadius: '0.5rem',
-      borderColor: '#d1d5db', // cinza claro
+      borderColor: '#e5e7eb',
       padding: '2px',
       boxShadow: 'none',
       "&:hover": {
-        borderColor: '#1f7e1f', // verde hover
+        borderColor: '#1f7e1f',
       },
     }),
   };
@@ -105,8 +105,8 @@ function ManageUsers() {
     ...theme,
     colors: {
       ...theme.colors,
-      primary25: '#d1fae5', // verde claro no hover
-      primary: '#165a16',   // verde principal
+      primary25: '#d1fae5',
+      primary: '#165a16',
     },
   });  
 
@@ -192,100 +192,155 @@ function ManageUsers() {
   };
 
   return (
-    <div className="min-h-screen w-full flex bg-[#eaeaea]"> 
+    <div className="min-h-screen w-full flex bg-gradient-to-br from-gray-50 to-gray-100"> 
       <CustomSidebar />
       <ToastContainer />
+      
       {/* Container principal */}
-      <div className="flex-1 flex items-center justify-center">
-        <div className="flex w-full max-w-6xl gap-8 justify-center">
-          {/* Formulário para adicionar usuário */}
-          <div className="bg-white/10 w-1/3 max-w-md backdrop-blur-sm rounded-lg p-8 shadow-lg">
-            <h2 className="text-2xl font-bold mb-4 text-center">Adicionar Usuário</h2>
-  
-            <form onSubmit={handleAddUser} className="space-y-4">
-              <div>
-                <label htmlFor="username" className="block text-sm font-medium mb-2">Nome</label>
-                <input
-                  type="text"
-                  id="username"
-                  name="username"
-                  value={newUser.username}
-                  onChange={handleInputChange}
-                  className="w-full p-2 border border-gray-300 rounded-lg"
-                />
-              </div>
-              <div>
-                <label htmlFor="password" className="block text-sm font-medium mb-2">Senha</label>
-                <input
-                  type="password"
-                  id="password"
-                  name="password"
-                  value={newUser.password}
-                  onChange={handleInputChange}
-                  className="w-full p-2 border border-gray-300 rounded-lg"
-                />
-              </div>
-              <div>
-                <label htmlFor="userType" className="block text-sm font-medium mb-2">Tipo de Usuário</label>
-                <Select
-                  instanceId={"userTypeSelect"}
-                  id="userType"
-                  name="userType"
-                  options={userTypeOptions}
-                  value={userTypeOptions.find(option => option.value === newUser.userType)}
-                  onChange={handleUserTypeChange}
-                  className="basic-single"
-                  classNamePrefix="select"
-                  styles={customStyles}
-                  theme={customTheme}
-                />
-              </div>
-              <button
-                type="submit"
-                className="mt-6 w-full bg-[#165a16] text-white py-2 px-4 rounded-lg hover:bg-[#1f7e1f] transition-all duration-400 ease-in-out"
-              >
-                Adicionar Usuário
-              </button>
-            </form>
+      <div className="flex-1 p-8">
+        <div className="max-w-7xl mx-auto">
+          {/* Cabeçalho da página */}
+          <div className="mb-8">
+            <h1 className="text-2xl font-bold mb-4 text-center">Gerenciamento de Usuários</h1>
+            <p className="text-center text-sm mb-6">Adicione, edite e gerencie os usuários do sistema</p>
           </div>
-          {/* Lista de usuários */}
-          <div className="bg-white/10 w-1/3 max-w-md backdrop-blur-sm rounded-lg p-8 shadow-lg">
-            <h2 className="text-xl font-bold mb-4">Usuários Cadastrados</h2>
-            <div className=""> 
-              {users.length > 0 ? (
-                <ul className="space-y-4">
-                  {users.map((user) => (
-                    <li key={user.id} className="border-b pb-2 last:border-b-0">
-                      <div className="flex justify-between items-center"> {/* Alinhando as informações à esquerda e os botões à direita */}
-                        <div>
-                          <p className="font-semibold">{user.username}</p>
-                          <p className="text-gray-500 text-sm">Tipo: {user.userType}</p>
+
+          <div className="grid grid-cols-1 lg:grid-cols-2 gap-8">
+            {/* Formulário para adicionar usuário */}
+            <div className="bg-gray-50 rounded-xl border border-gray-300 p-6">
+              <div className="flex items-center mb-6">
+                <div className="w-8 h-8 bg-green-100 rounded-lg flex items-center justify-center mr-3">
+                  <svg className="w-5 h-5 text-green-600" fill="none" stroke="currentColor" viewBox="0 0 24 24">
+                    <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M12 6v6m0 0v6m0-6h6m-6 0H6" />
+                  </svg>
+                </div>
+                <h2 className="text-xl font-semibold text-gray-800">Adicionar Novo Usuário</h2>
+              </div>
+  
+              <form onSubmit={handleAddUser} className="space-y-5">
+                <div>
+                  <label htmlFor="username" className="block text-sm font-medium text-gray-700 mb-2">
+                    Nome do Usuário
+                  </label>
+                  <input
+                    type="text"
+                    id="username"
+                    name="username"
+                    value={newUser.username}
+                    onChange={handleInputChange}
+                    className="w-full p-3 border border-gray-300 rounded-lg focus:ring-2 focus:ring-green-500 focus:border-transparent transition-all duration-200"
+                    placeholder="Digite o nome do usuário"
+                  />
+                </div>
+                
+                <div>
+                  <label htmlFor="password" className="block text-sm font-medium text-gray-700 mb-2">
+                    Senha
+                  </label>
+                  <input
+                    type="password"
+                    id="password"
+                    name="password"
+                    value={newUser.password}
+                    onChange={handleInputChange}
+                    className="w-full p-3 border border-gray-300 rounded-lg focus:ring-2 focus:ring-green-500 focus:border-transparent transition-all duration-200"
+                    placeholder="Digite a senha"
+                  />
+                </div>
+                
+                <div>
+                  <label htmlFor="userType" className="block text-sm font-medium text-gray-700 mb-2">
+                    Tipo de Usuário
+                  </label>
+                  <Select
+                    instanceId={"userTypeSelect"}
+                    id="userType"
+                    name="userType"
+                    options={userTypeOptions}
+                    value={userTypeOptions.find(option => option.value === newUser.userType)}
+                    onChange={handleUserTypeChange}
+                    className="basic-single"
+                    classNamePrefix="select"
+                    styles={customStyles}
+                    theme={customTheme}
+                    placeholder="Selecione o tipo de usuário"
+                  />
+                </div>
+                
+                <button
+                  type="submit"
+                  className="w-full bg-[#165a16] text-white py-2 px-4 rounded-lg hover:bg-[#208120] transition-all duration-300 ease-in-out"
+                >
+                  Adicionar Usuário
+                </button>
+              </form>
+            </div>
+
+            {/* Lista de usuários */}
+            <div className="bg-[gray-50] rounded-xl border border-gray-300 p-6">
+              <div className="flex items-center mb-6">
+                <div className="w-8 h-8 bg-green-100 rounded-lg flex items-center justify-center mr-3">
+                  <svg className="w-5 h-5 text-green-600" fill="none" stroke="currentColor" viewBox="0 0 24 24">
+                    <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M17 20h5v-2a3 3 0 00-5.356-1.857M17 20H7m10 0v-2c0-.656-.126-1.283-.356-1.857M7 20H2v-2a3 3 0 015.356-1.857M7 20v-2c0-.656.126-1.283.356-1.857m0 0a5.002 5.002 0 019.288 0M15 7a3 3 0 11-6 0 3 3 0 016 0zm6 3a2 2 0 11-4 0 2 2 0 014 0zM7 10a2 2 0 11-4 0 2 2 0 014 0z" />
+                  </svg>
+                </div>
+                <h2 className="text-xl font-semibold text-gray-800">Usuários Cadastrados</h2>
+                <span className="ml-auto bg-green-100 text-green-700 px-3 py-1 rounded-full text-sm font-medium">
+                  {users.length} usuário{users.length !== 1 ? 's' : ''}
+                </span>
+              </div>
+              
+              <div className="h-96 overflow-y-auto pr-2 space-y-3">
+                {users.length > 0 ? (
+                  users.map((user) => (
+                    <div key={user.id} className="border border-gray-200 rounded-lg p-4 hover:bg-gray-50 transition-all duration-200">
+                      <div className="flex items-center justify-between">
+                        <div className="flex-1">
+                          <div className="flex items-center mb-1">
+                            <h3 className="font-semibold text-gray-800 mr-3">{user.username}</h3>
+                            <span className={`px-2 py-1 rounded-full text-xs font-medium ${
+                              user.userType === 'admin' 
+                                ? 'bg-purple-100 text-purple-800' 
+                                : 'bg-green-100 text-green-800'
+                            }`}>
+                              {user.userType === 'admin' ? 'Administrador' : 'Produtor'}
+                            </span>
+                          </div>
+                          <p className="text-sm text-gray-500">ID: {user.id}</p>
                         </div>
-                        <div className="flex space-x-4"> {/* Alinhando os botões ao lado */}
+                        
+                        <div className="flex space-x-2">
                           <button
                             onClick={() =>
                               handleUpdateUserType(user.id, user.userType === "prod" ? "admin" : "prod")
                             }
-                            className="bg-[#165a16] hover:bg-[#1f7e1f] rounded-sm text-sm px-4 py-2 text-center text-white transition-all duration-300 ease-in-out"
+                            className="bg-green-600 hover:bg-green-700 text-white px-3 py-2 rounded-lg text-sm font-medium transition-all duration-200"
                           >
                             Alterar
                           </button>
                           <button
                             onClick={() => handleDeleteUser(user.id)}
-                            className="bg-red-600 hover:bg-red-800 rounded-sm text-sm px-4 py-2 text-center text-white transition-all duration-300 ease-in-out"
+                            className="bg-red-600 hover:bg-red-700 text-white px-3 py-2 rounded-lg text-sm font-medium transition-all duration-200"
                           >
                             Excluir
                           </button>
                         </div>
                       </div>
-                    </li>
-                  ))}
-                </ul>
-              ) : (
-                <p className="text-gray-600">Nenhum usuário cadastrado.</p>
-              )}
+                    </div>
+                  ))
+                ) : (
+                  <div className="text-center py-8">
+                    <div className="w-16 h-16 bg-green-100 rounded-full flex items-center justify-center mx-auto mb-4">
+                      <svg className="w-8 h-8 text-green-600" fill="none" stroke="currentColor" viewBox="0 0 24 24">
+                        <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M17 20h5v-2a3 3 0 00-5.356-1.857M17 20H7m10 0v-2c0-.656-.126-1.283-.356-1.857M7 20H2v-2a3 3 0 015.356-1.857M7 20v-2c0-.656.126-1.283.356-1.857m0 0a5.002 5.002 0 019.288 0M15 7a3 3 0 11-6 0 3 3 0 016 0zm6 3a2 2 0 11-4 0 2 2 0 014 0zM7 10a2 2 0 11-4 0 2 2 0 014 0z" />
+                      </svg>
+                    </div>
+                    <p className="text-gray-500 font-medium">Nenhum usuário cadastrado</p>
+                    <p className="text-gray-400 text-sm">Adicione o primeiro usuário usando o formulário ao lado</p>
+                  </div>
+                )}
+              </div>
             </div>
-
           </div>
         </div>
       </div>
